Application improvements

+++ src/sampletones/data/types.py
@@ -3,6 +3,6 @@ from enum import StrEnum
 
 class DataType(StrEnum):
     UOFFSET = "uoffset"
-    FLOAT64 = "float64"
+    FLOAT32 = "float32"
     INT32 = "int32"
     BOOL = "bool"
